Implement low-level software trace recording, in C, in libcore.
This is implemented in C because if it were implemented in Rust, then
the tracer code would trace itself and everything it calls (or we'd end
up annotating a lot of the standard library `#[no_trace]`).

With each call to the trace recorder, we build a "MIR location" struct
from the call arguments and insert this into a thread local trace
buffer.

A successful tracing session will return the trace buffer to the
consumer when we stop tracing. If however something happens which
renders the trace data unreliable, we say the trace was "invalidated"
and return NULL.

For example, a trace is invalidated when the Rust stack overflow
detector's signal handler interrupts a trace. A trace is also
invalidated when when the trace buffer can't be a allocated or
re-allocated.

User signal handlers remain an open problem, as we are unable to force
code outside of our hands to invalidate a trace.