The beginnings of a new "TIR" serialiser.
This replaces the custom MIR serialiser, which a serde-based TIR
(tracing IR) serialiser. The change of name is to signify that we are
actually implementing our own IR now, not just a mirror of MIR.

The conversion uses a trait `ToPack` which is effectively a poor man's
`From` trait. We couldn't use the `From` trait since the "from" and "to"
data structures are both external crates in the eyes of the compiler,
and Rust currently disallows this.

While we are here, remove the "YK_DEBUG_SECTIONS" environment variable
switch, as we are always going to need this information. It's not just a
debugging aid.

We do not yet implement MIR to TIR statement translation. This will be
started in my next PR.

I plan to rename the file `mir_cfg.rs` to something more appropriate
later (say `gen_tir.rs`), but only once my outstanding branches have
been merged.