Start converting MIR to SSA TIR.
This change:

 - Starts adding support for statements to TIR.
   Only simple assignments are translated. Everything else is made into
   an "unimplemented" statement.

 - Does the first part of the SSA conversion, namely:
   * Computes the dominance frontier.
   * Inserts PHI nodes in the right place and for the correct TIR variables.

So what we get is a CFG with PHIs, but which is not yet in SSA. The
remaining step is to walk the dominator tree and assign each definition
of a variable with a new SSA variable name.

Manual inspection of CFG graphs with assignments and PHIs shown
look good, but tests will follow later.