TIR test suite.
This change adds a rustc flag `--emit yk-tir` which causes ykrustc to
dump the TIR in a textual format and exit.

Then we add a test new test suite kind which allows us to write tests
against the textual TIR dumps. It works in much the same way as the
existing MirOpt test suite. In fact, we re-use a lot of that code.