Implement and use a query to help us find MIR in dependent crates.
The new query returns a set of DefIds which:
 * Have an entry in the definition path table in the metadata.
 * Have MIR available.

At serialisation time, we then iterate over all dependent crates, adding
new DefIds into the list of things to be serialized. All items in the
local crate and dependent crates are serialised, as it's hard to know
what code will actually be executed at runtime (thanks to dynamic
dispatch).

Sadly this makes testing the compiler very slow and it also uses quite a
lot of RAM.

I theorise that It's slow because, for each binary we build (of which
there are many in tests) we serialise all dependent crates to SIR in the
resulting ELF binary. I think We could improve test times by requiring a
compiler flag or environment variable to be set when you are building
code that you actually want to trace (but which is unset for testing).

I theorise that it uses a lot of RAM because to serialise the SIR we
build a large DefIdSet for the local crate and dependent crates. Instead
of having one large SIR section, we could have smaller ones for each
crate, then we wouldn't need the huge DefIdSet in memory all at once.