Replace YK_TRACER with a proper tracked compiler flag.
By using a tracked flag, it means that if the tracer choice changes, a
rebuild is triggered.

We still need an environment variable to choose what kind of standard
library to build at bootstrap time (`STD_TRACER_MODE`), but its role is
minimal. For now we hard-code `STD_TRACER_MODE=sw`, but we plan to have
CI test both modes. At that time, the CI will set this environment
variable instead of hard-coding.

We also get rid of "sw-nosir" mode, as there's a section of code in the
bootstrapper which allows us to apply flags for the standard library
only. We now use this to turn on a tracer for the standard lib only when
building the compiler, thus not slowing down the tests (this was the
reason we had 'sw-nosir' in the first place).