Add YK labels when returning from a function call
The MIR level and the binary level handle function calls that return a
value differently. The former treats the assignment of that value as
part of the function block, whereas the latter treats it as part of the
block following the return. Although, this doesn't make a difference
when running programs, it means that the block label for the following
block is at the wrong location, which makes it impossible to assign it
to the correct trace.

However, we can't simply adjust the label by x bytes, since this problem
only occurs when the result of the function call is assigned to a
variable. We thus add another label `YK_RET` which is inserted during a
function call, just before the return value is assigned. This places
this new label at the correct location and allows us to match it to the
same location in our trace.