Cache the results of the monomorphisation call resolver.
To cut a long story short, in order to determine static call targets
(where possible) it is necessary to call Instance::resolve(). However,
only the monomorphisation stage is permitted to do so.

This change caches the results of resolutions so that we can use them
during SIR lowering.

We also utilise the new CallOperand::Virtual variant for calls which
cannot be statically known at compile time.

Discussion here:
https://rust-lang.zulipchat.com/#narrow/stream/182449-t-compiler.2Fhelp/topic/Calls.20in.20MIR