Move SIR generation one level down into the LLVM codegen.
This allows us to serialise post-monomorphisation and will solve a
problem we had before: we were unable to resolve the
pre-monomorphisation call targets.

In doing so, we found a better way to link ELF objects, so
librustc_yk_link and librustc_yk_sections are entirely removed.

We currently only serialise functions with empty blocks. Populating
those is future work.

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>