Start work on local variables and instructions.
Starting with store and load, as these are among the simplest.

Store instructions are simple, as they don't generate a value, but load
requires a little leg-work:

The way LLVM API works, when an instruction which generates a result is
built, an implicit assignment is created behind the scenes.

For example, to emit a load from C, you'd do: `LLVMBuildLoad(x, "res")`,
where `x` is an LLVM pointer representing where to load from, and  "res"
is the name of the result. This is equivalent to an LLVM bitcode
statement `%res = load(%x)`. Put differently, each computation adds a
new local variable to the function you are emitting instructions into.

Further, the call to `LLVMBuildLoad` returns another LLVM pointer
representing the variable which will store the result of the load. To
use the result as an operand for a further computation, you'd pass the
this pointer as an argument to another API call.

E.g.
```
LLVMValueRef *a = LLVMBuildLoad(x, "a");
LLVMValueRef *b = LLVMBuildLoad(y, "y");
LLVMValueRef *c = LLVmBuildAdd(a, b, "c");
```

Note that each ValueRef in the above also identifies the instruction
which generates the value.

The upshot of all of this is that in order to build SIR instructions
from within the LLVM codegen we need to be able to identify the results
of SIR computations from LLVM ValueRef pointers.

To this end, we add a `Local` variant to `sir::Value` and when a `load`
instruction is emitted, we create a SIR load statement and a SIR local
variable for the result, but also cache the LLVM pointer.

This mechanism will be used for all instructions which generate a value,
not just for loads.

Future work: Currently every load operand is assumed to be a local
variable, but it could be something else, like a constant, or a static.