SIRv3.
This change moves SIR generation from librustc_codegen_llvm to
librustc_codegen_ssa.

This means that SIR will be generated from *monomorphised* MIR, rather
than from LLVM API calls. This is easier to implement and makes it
easier to access high-level program information, but means that our
notion of a block is a little further away from what ends up in the
binary. We've discussed this and think that it probably isn't going to
be an issue.

For now all SIR statements and terminators are dummy "unimplemented"
place-holders. Next we will start serialising them properly.

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>