Lower MIR call terminators to their SIR equivalent.
The target function must be monomorphised. We had two options:

 - Duplicate monomorphisation logic in codegen_block().

 - Lower terminators deeper in the codegen and reuse existing
   monomorphisation logic.

I don't like the idea of having to keep monomorphisation logic in-sync
with upstream, so I've opted for the latter. This strategy means we now
have lots of places where we have to lower a terminator, whereas before
we had only one.

I can live with this, however I did add a macro to help us identify the
many (currently) unimplemented terminator insertions.