Make no attempt to build a traceable std library (for now).
To (reliably) trace through the std lib, we need to build it unoptimised
(to stop LLVM re-ordering blocks), but it's currently not possible to do
so.

https://rust-lang.zulipchat.com/#narrow/stream/182449-t-compiler.2Fhelp/topic/Unoptimised.20compiler.20builds

For now we will trace only user code and come back to this problem
later.

The good news is that a lot of the std lib is parametrized or inlined,
so compilation of those functions will be delayed until compiling the
user program. In other words, we should still be able to trace those.