Add back a tracer selection flag (and associated cfg).
This adds a codegen flag `-C yk_tracer=T`, where (for now) T is either
`hw` (hardware tracing) or `off`. This in turn allows compile-time cfg
tests like `#[cfg(yk_tracer = "hw")]`.

The `yk_` prefix is deliberate. In the last incarnation of Yorick, this
flag was `-C tracer=T`. I'd like to "namespace" all of the
Yorick-related bits in the rust compiler so that it's obvious what is to
do with our JIT and what isn't.

Also before the `-C tracer` flag set the `tracermode` cfg, which seemed
inconsistent. Now both the name of the flag and the cfg we set are both
`yk_tracer`.